requirements.txt
numpy
Biopython
scipy
tqdm
h5py
joblib